final/requirements.txt
aiohttp==3.9.1
aiofiles==23.2.1
asyncio==3.4.3
celery==5.3.6
python-dotenv==1.0.0
SQLAlchemy==2.0.23
redis==5.0.1
gunicorn==21.2.0